// frontend/src/types/apps/initiativeTypes.ts
export type InitiativeType = {
    initiatives_id: number
    sr_no: string
    definition: string
    user_id: number
    s_o_id: number
    strategic_objectives_description: string
    dept_id: number
    dept_name: string
    section_id: number
    section_name: string
    status_id: number
    status_name: string
    percentage: string
    start_date: string
    end_date: string
    comment?: string
}